Allow for Client Jar Caching (#66)
This will allow for the client .jar file to be cached by power uses (normal users aren't impacted). 

This will impact anyone who uses a custom configuration which has the "texturepath" line in it. It's likely that if a user has a custom config, they have that option set since we used to have the .jar in a non-default location. For those users, we display a warning and copy the .jar into the old locations.

In the future (July) we will remove the workaround and force users to update their configurations.